requirements.txt
mysqlclient
Django
python-dotenv
django-bootstrap-v5
django-ckeditor
